requirements.txt
# Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Base de datos
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Variables de entorno
python-dotenv==1.0.0

# Machine Learning / YOLO
ultralytics==8.0.207
opencv-python==4.9.0.80
numpy==1.24.3

# AWS S3
boto3==1.34.51

# OpenAI
openai==1.12.0

# Reportes PDF
reportlab==4.0.9

# Utilidades
python-multipart==0.0.6
requests==2.31.0
pillow==10.2.0